backend/requirements.txt
altair==5.5.0
asgiref==3.11.1
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
executing==2.2.1
flatbuffers==25.2.10
gitdb==4.0.12
GitPython==3.1.44
gunicorn==25.1.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.9.1
lazy_loader==0.4
llvmlite==0.45.0
MarkupSafe==3.0.2
matplotlib-inline==0.2.1
mpmath==1.3.0
narwhals==1.43.1
nest-asyncio==1.6.0
networkx==3.5
numba==0.62.0
numpy==2.2.6
onnxruntime==1.22.1
onnxruntime-gpu==1.22.0
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.0
parso==0.8.5
pillow==11.2.1
platformdirs==4.4.0
pooch==1.8.2
prompt_toolkit==3.0.52
protobuf==6.31.1
psutil==7.1.3
pure_eval==0.2.3
pyarrow==20.0.0
pydeck==0.9.1
Pygments==2.19.2
PyMatting==1.1.14
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
referencing==0.36.2
rembg==2.0.67
requests==2.32.4
rpds-py==0.25.1
scikit-image==0.25.2
scipy==1.16.2
six==1.17.0
smmap==5.0.2
sqlparse==0.5.5
stack-data==0.6.3
streamlit==1.46.0
sympy==1.14.0
tenacity==9.1.2
tifffile==2025.9.20
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.14
